% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spectra.R
\name{mod_spectraUI}
\alias{mod_spectraUI}
\title{Spectrum Viewer Module UI}
\usage{
mod_spectraUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A \code{\link[plotly]{plotlyOutput}} element.
}
\description{
Renders an interactive \code{\link[gutenTAG]{plotMetapeaks}} plot. Users
  click within a metapeak region to select that channel for the image panel.
}
\examples{
if (interactive()) {
  ui <- shiny::fluidPage(mod_spectraUI("spectra"))
  shiny::shinyApp(ui, function(input, output, session) {})
}
}
